Added Quote & http client DI for Chia rpc client

+++ SwapHunter/Client/IChiaRpcClient.cs
@@ -9,6 +9,6 @@ namespace SwapHunter.Client
 {
   public interface IChiaRpcClient
   {
-    Task<CreateOfferResponse> CreateOffer(string assetid, int requesting_amount, int xch_amount, int fee);
+    Task<CreateOfferResponse> CreateOffer(string assetid, double requesting_amount, double xch_amount, double fee);
   }
 }
